refactor: extract release plan step builder
